Simplify regex for shell_functions.
Additionally, this prevents `_alias_tips__preexec` from getting called
twice which saves about 100ms on startup.  I'm not sure why it has this
effect, but I suspect it has something to do with quoting and the dollar
sign in the original regex.

Here's the zprof info from config before this commit:

|          | Calls |  Total |   Each |
|----------+-------+--------+--------|
| Original |     2 | 215.33 | 107.67 |
| New      |     1 | 108.31 | 108.31 |